refactor: move personal images to separate component
